Review: move RegisterIterator to register.rs, priv RegisterIterator::new(), newline post Register::values()
