+++ src/rust/iced-x86/src/lib.rs
@@ -1273,7 +1273,6 @@ mod memory_size;
 mod mnemonic;
 mod mnemonics;
 mod register;
-mod register_iterator;
 #[cfg(test)]
 pub(crate) mod test;
 #[cfg(test)]
@@ -1301,4 +1300,3 @@ pub use self::instruction::*;
 pub use self::memory_size::*;
 pub use self::mnemonic::*;
 pub use self::register::*;
-pub use self::register_iterator::*;
